+++ nilearn/_utils/tests/test_glm.py
@@ -1,0 +1,17 @@
+import pandas as pd
+import pytest
+
+from nilearn._utils.glm import check_and_load_tables
+
+
+def test_img_table_checks():
+    # check tables type and that can be loaded
+    with pytest.raises(ValueError, match="table path .* could not be loaded"):
+        check_and_load_tables([".csv", ".csv"], "")
+    with pytest.raises(
+        TypeError,
+        match="can only be a pandas DataFrame, a Path object or a string",
+    ):
+        check_and_load_tables([[], pd.DataFrame()], "")
+    with pytest.raises(ValueError, match="table path .* could not be loaded"):
+        check_and_load_tables([".csv", pd.DataFrame()], "")
