+++ nilearn/glm/tests/test_first_level.py
@@ -44,7 +44,6 @@ from nilearn.glm.first_level.first_level import (
     _check_trial_type,
     _list_valid_subjects,
     _yule_walker,
-    check_and_load_tables,
 )
 from nilearn.glm.regression import ARModel, OLSModel
 from nilearn.image import get_data
@@ -2153,17 +2152,6 @@ def test_img_table_checks():
     # check matching lengths
     with pytest.raises(ValueError, match="len.* does not match len.*"):
         _check_list_length_match([""] * 2, [""], "", "")
-
-    # check tables type and that can be loaded
-    with pytest.raises(ValueError, match="table path .* could not be loaded"):
-        check_and_load_tables([".csv", ".csv"], "")
-    with pytest.raises(
-        TypeError,
-        match="can only be a pandas DataFrame, a Path object or a string",
-    ):
-        check_and_load_tables([[], pd.DataFrame()], "")
-    with pytest.raises(ValueError, match="table path .* could not be loaded"):
-        check_and_load_tables([".csv", pd.DataFrame()], "")
 
 
 # -----------------------surface tests--------------------------------------- #
